Hg-commit: e4f9833c32bdffd114d2089ae117651b77159d06

+++ config/builder.py
@@ -93,6 +93,7 @@ regressionParameters = {'src/sys/comm/examples/tests/ex1':    [{'numProcs': 2},
 -snes_mf_operator -pack_dm_mat_type aij -pc_type fieldsplit -pc_fieldsplit_type additive -fieldsplit_u_ksp_type gmres -fieldsplit_k_pc_type jacobi'},
                                                                {'numProcs': 1, 'args': '-da_grid_x 20 -snes_converged_reason -snes_monitor_short -ksp_monitor_short -problem_type 2 \
 -snes_mf_operator -pack_dm_mat_type nest -pc_type fieldsplit -pc_fieldsplit_type additive -fieldsplit_u_ksp_type gmres -fieldsplit_k_pc_type jacobi'}],
+                        'src/snes/examples/tutorials/ex33':   [{'numProcs': 1, 'args': '-snes_converged_reason -snes_monitor_short'}],
                         'src/snes/examples/tutorials/ex52':   [{'numProcs': 1, 'args': '-dm_view -refinement_limit 0.0625 -compute_function -batch',
                                                                 'setup': './bin/pythonscripts/PetscGenerateFEMQuadrature.py 2 1 src/snes/examples/tutorials/ex52.h'}],
                         'src/ts/examples/tutorials/ex18':      {'numProcs': 1, 'args': '-snes_mf -ts_monitor_solution -ts_monitor -snes_monitor'},
@@ -1223,7 +1224,8 @@ class PETScMaker(script.Script):
    if os.path.isfile(self.sourceDBFilename):
      os.remove(self.sourceDBFilename)
      self.logPrint('Removed '+self.sourceDBFilename)
-   shutil.rmtree(self.getObjDir(libname))
+   if os.path.exists(self.getObjDir(libname)):
+     shutil.rmtree(self.getObjDir(libname))
    self.logPrint('Removed '+self.getObjDir(libname))
    return
